refact(modularization/hosts): categorizing hosts as module
