+++ dashboard/src/modules/hosts/dialogs/mutation/fields/fingerprint.tsx
