+++ dashboard/src/modules/hosts/dialogs/mutation/fields/index.ts
