+++ dashboard/src/modules/hosts/dialogs/mutation/fields/remark.tsx
@@ -9,7 +9,7 @@ import {
     PopoverTrigger,
     PopoverContent,
 } from "@marzneshin/components";
-import { PopoverGuide } from '@marzneshin/features/hosts';
+import { PopoverGuide } from '@marzneshin/modules/hosts';
 import { useFormContext } from "react-hook-form";
 import { useTranslation } from "react-i18next";
 import { InfoIcon } from 'lucide-react';
