+++ dashboard/src/modules/hosts/dialogs/mutation/fields/weight.tsx
