+++ dashboard/src/modules/hosts/index.ts
