+++ dashboard/src/modules/hosts/schema/get-default.ts
