+++ dashboard/src/modules/hosts/schema/hosts-schema.ts
