+++ dashboard/src/modules/hosts/schema/index.ts
