+++ dashboard/src/modules/hosts/services/host-mutation.dto.ts
@@ -1,4 +1,4 @@
-import { HostType } from "@marzneshin/features/hosts";
+import { HostType } from "@marzneshin/modules/hosts";
 
 export interface HostRequestDto {
     inboundId: number
