+++ dashboard/src/modules/hosts/services/host.query.ts
@@ -1,4 +1,4 @@
-import { HostType } from "@marzneshin/features/hosts";
+import { HostType } from "@marzneshin/modules/hosts";
 import { useQuery } from "@tanstack/react-query";
 import { fetch } from "@marzneshin/utils";
