+++ dashboard/src/modules/hosts/services/inbound.query.ts
@@ -1,4 +1,4 @@
-import { InboundType } from "@marzneshin/features/inbounds";
+import { InboundType } from "@marzneshin/modules/inbounds";
 import { useQuery } from "@tanstack/react-query";
 import { fetch } from "@marzneshin/utils";
