+++ dashboard/src/modules/hosts/services/update-host.mutate.ts
@@ -1,4 +1,4 @@
-import { HostType } from "@marzneshin/features/hosts";
+import { HostType } from "@marzneshin/modules/hosts";
 import { useMutation } from "@tanstack/react-query";
 import { fetch, queryClient } from "@marzneshin/utils";
 import { toast } from "sonner";
