+++ dashboard/src/modules/hosts/tables/index.ts
