+++ dashboard/src/modules/hosts/types/index.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { HostSchema } from "@marzneshin/features/hosts";
+import { HostSchema } from "@marzneshin/modules/hosts";
 
 export type HostSchemaType = z.infer<typeof HostSchema>;
 export type HostType = HostSchemaType & { id?: number };
