+++ dashboard/src/routes/_dashboard/hosts/$hostId/index.tsx
@@ -5,7 +5,7 @@ import {
 import {
     useRouterHostContext,
     HostSettingsDialog,
-} from "@marzneshin/features/hosts";
+} from "@marzneshin/modules/hosts";
 import { useDialog } from "@marzneshin/hooks";
 
 const HostOpen = () => {
